Disable rsync of non-existent ftp/ directory for arm.
Sponsored by:	The FreeBSD Foundation

git-svn-id: file:///Users/gjb/Downloads/releng-svn/thermite-r688@138 5711f056-dcf5-4cc1-a278-612e92bcda7e

+++ user/gjb/scripts/release/ftp-stage.sh
@@ -123,9 +123,11 @@ setup_stageenv() {
 }
 
 stage_isos() {
-	echo "=== Rsync ${C}/R/ftp to ${_ftpdir}/${path}/${OSRELEASE}..."
-	rsync -a --delete ${C}/R/ftp/* \
-		${_ftpdir}/${path}/${OSRELEASE}/
+	if [ "X${arch}" != "Xarmv6" ]; then
+		echo "=== Rsync ${C}/R/ftp to ${_ftpdir}/${path}/${OSRELEASE}..."
+		rsync -a --delete ${C}/R/ftp/* \
+			${_ftpdir}/${path}/${OSRELEASE}/
+	fi
 
 	# FreeBSD-11.0-CURRENT-amd64
 	oldname="${__DISCNAME}"
